Refactor GUI dialogs into separate modules
Moved dialog classes from a monolithic dialogs.py file into individual modules under src/gambitpairing/gui/dialogs/. Updated imports and re-exports for backward compatibility. Also fixed PyQt import in printing.py.
