+++ src/gambitpairing/gui/dialogs/settings_dialog.py
@@ -1,0 +1,91 @@
+from PyQt6 import QtWidgets
+from PyQt6.QtCore import Qt
+from typing import List, Tuple
+from gambitpairing.core.constants import TIEBREAK_NAMES
+
+
+class SettingsDialog(QtWidgets.QDialog):
+    def __init__(self, num_rounds: int, tiebreak_order: List[str], parent=None):
+        super().__init__(parent)
+        self.setWindowTitle("Tournament Settings")
+        self.setMinimumWidth(350)
+        self.current_tiebreak_order = list(tiebreak_order)
+        layout = QtWidgets.QVBoxLayout(self)
+        rounds_group = QtWidgets.QGroupBox("General")
+        rounds_layout = QtWidgets.QFormLayout(rounds_group)
+        self.spin_num_rounds = QtWidgets.QSpinBox()
+        self.spin_num_rounds.setRange(1, 50)
+        self.spin_num_rounds.setValue(num_rounds)
+        self.spin_num_rounds.setToolTip("Set the total number of rounds.")
+        rounds_layout.addRow("Number of Rounds:", self.spin_num_rounds)
+        layout.addWidget(rounds_group)
+        tiebreak_group = QtWidgets.QGroupBox("Tiebreak Order")
+        tiebreak_layout = QtWidgets.QHBoxLayout(tiebreak_group)
+        self.tiebreak_list = QtWidgets.QListWidget()
+        self.tiebreak_list.setToolTip(
+            "Order in which tiebreaks are applied (higher is better). Drag to reorder."
+        )
+        self.tiebreak_list.setDragDropMode(
+            QtWidgets.QAbstractItemView.DragDropMode.InternalMove
+        )
+        self.populate_tiebreak_list()
+        tiebreak_layout.addWidget(self.tiebreak_list)
+        move_button_layout = QtWidgets.QVBoxLayout()
+        btn_up = QtWidgets.QPushButton("Up")
+        btn_down = QtWidgets.QPushButton("Down")
+        btn_up.clicked.connect(self.move_tiebreak_up)
+        btn_down.clicked.connect(self.move_tiebreak_down)
+        move_button_layout.addStretch()
+        move_button_layout.addWidget(btn_up)
+        move_button_layout.addWidget(btn_down)
+        move_button_layout.addStretch()
+        tiebreak_layout.addLayout(move_button_layout)
+        layout.addWidget(tiebreak_group)
+        self.buttons = QtWidgets.QDialogButtonBox(
+            QtWidgets.QDialogButtonBox.StandardButton.Ok
+            | QtWidgets.QDialogButtonBox.StandardButton.Cancel
+        )
+        self.buttons.setSizePolicy(
+            QtWidgets.QSizePolicy.Policy.Expanding, QtWidgets.QSizePolicy.Policy.Minimum
+        )
+        self.buttons.setMinimumHeight(40)
+        self.buttons.accepted.connect(self.accept)
+        self.buttons.rejected.connect(self.reject)
+        layout.addWidget(self.buttons)
+
+    def populate_tiebreak_list(self):
+        self.tiebreak_list.clear()
+        for tb_key in self.current_tiebreak_order:
+            display_name = TIEBREAK_NAMES.get(tb_key, tb_key)
+            item = QtWidgets.QListWidgetItem(display_name)
+            item.setData(Qt.ItemDataRole.UserRole, tb_key)
+            self.tiebreak_list.addItem(item)
+
+    def move_tiebreak_up(self):
+        current_row = self.tiebreak_list.currentRow()
+        if current_row > 0:
+            item = self.tiebreak_list.takeItem(current_row)
+            self.tiebreak_list.insertItem(current_row - 1, item)
+            self.tiebreak_list.setCurrentRow(current_row - 1)
+            self.update_order_from_list()
+
+    def move_tiebreak_down(self):
+        current_row = self.tiebreak_list.currentRow()
+        if current_row < self.tiebreak_list.count() - 1:
+            item = self.tiebreak_list.takeItem(current_row)
+            self.tiebreak_list.insertItem(current_row + 1, item)
+            self.tiebreak_list.setCurrentRow(current_row + 1)
+            self.update_order_from_list()
+
+    def update_order_from_list(self):
+        self.current_tiebreak_order = [
+            self.tiebreak_list.item(i).data(Qt.ItemDataRole.UserRole)
+            for i in range(self.tiebreak_list.count())
+        ]
+
+    def accept(self):
+        self.update_order_from_list()
+        super().accept()
+
+    def get_settings(self) -> Tuple[int, List[str]]:
+        return self.spin_num_rounds.value(), self.current_tiebreak_order
