+++ src/gambitpairing/gui/dialogs/update_prompt_dialog.py
