+++ src/gambitpairing/gui/notification.py
@@ -1,5 +1,5 @@
 from PyQt6 import QtWidgets, QtCore, QtGui
-from core.utils import apply_stylesheet
+from gambitpairing.core.utils import apply_stylesheet
 
 def show_notification(parent, message: str, duration: int = 1500):
     """Show a floating notification label with fade-in/out animation over the parent widget."""
